+++ NEWS.md
@@ -2,6 +2,7 @@
 'replyr' 0.2.5 2017-03-24
 
  * Remove all use of dplyr underbar/underscore forms to future proof as these are already deprecated in dev ( https://github.com/hadley/dplyr/blob/master/R/manip.r 33d0b1ecaa22644e0d78345b13d5f8aca5d949b ).
+ * re-introduce moving values between rows and columns.
 
 'replyr' 0.2.4 2017-03-13